backend-app/pyproject.toml
[tool.poetry]
name = "algokit-deps-viz"
version = "0.1.0"
description = ""
authors = ["David Rojas <lempira@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
fastapi = {extras = ["standard"], version = "^0.115.6"}
pydantic-settings = "^2.7.1"
google-cloud-secret-manager = "^2.22.0"
google-cloud-storage = "^2.19.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
